refactor: LottoApp의 유효성검사 로직을 LottoValidator 클래스로 분리
